+++ server/src/Service/Messages/StreamResponse.cs
@@ -3,11 +3,11 @@ namespace KRPC.Service.Messages
     #pragma warning disable 1591
     public class StreamResponse : IMessage
     {
-        public uint Id { get; private set; }
+        public ulong Id { get; private set; }
 
         public Response Response { get; set; }
 
-        public StreamResponse (uint id)
+        public StreamResponse (ulong id)
         {
             Id = id;
         }
